feat(dto): Normalize pointer type conversion to concrete values

+++ internal/dto/v1/user_dto.go
@@ -21,10 +21,10 @@ type UserDTO struct {
 type CreateUserInput struct {
 	Name     string `json:"name" binding:"required"`
 	Email    string `json:"email" binding:"required,email"`
-	Age      int    `json:"age" binding:"omitempty,gt=18"`
+	Age      int32  `json:"age" binding:"omitempty,gt=18"`
 	Password string `json:"password" binding:"required,min=8,password_strong"`
-	Status   int    `json:"status" binding:"required,oneof=1 2 3"`
-	Level    int    `json:"level" binding:"required,oneof=1 2 3"`
+	Status   int32  `json:"status" binding:"required,oneof=1 2 3"`
+	Level    int32  `json:"level" binding:"required,oneof=1 2 3"`
 }
 
 type UpdateUserInput struct {
@@ -42,8 +42,8 @@ func (input *CreateUserInput) MapCreateInputToModel() sqlc.CreateUserParams {
 		UserPassword: input.Password,
 		UserFullname: input.Name,
 		UserAge:      utils.ConvertToInt32Pointer(input.Age),
-		UserStatus:   int32(input.Status),
-		UserLevel:    int32(input.Level),
+		UserStatus:   input.Status,
+		UserLevel:    input.Level,
 	}
 }
